1, set recv and send len type from uint16 to uint32, let app layer to choose the real type 2, implement peer function in starnet, not tested 3, add some function to starnet.h, for peer issue
vet/app: me

+++ StarNet/handler.h
@@ -6,7 +6,7 @@
 
 typedef bool (CALLBACK* ON_CONNECTION)(ConnID);
 typedef void (CALLBACK* ON_DISCONNECT)(ConnID);
-typedef void (CALLBACK* ON_DATA)(ConnID, uint16, char*);
+typedef void (CALLBACK* ON_DATA)(ConnID, uint32, char*);
 typedef void (CALLBACK* ON_CONNECT_FAILED)(void*);
 
 struct Handler
